+++ includes/ui/tabs/diagnostics.php
@@ -9,6 +9,15 @@ defined( '\\ABSPATH' ) || exit;
 
 ?>
 
-<p>
-    <textarea class="large-text readonly" rows="20" readonly><?php require __DIR__ . '/../../diagnostics.php'; ?></textarea>
+<div class="card">
+    <pre id="redis-cache-diagnostics"><?php require __DIR__ . '/../../diagnostics.php'; ?></pre>
+</div>
+
+<p id="redis-cache-copy-button">
+    <span class="copy-button-wrapper">
+        <button type="button" class="button copy-button" data-clipboard-target="#redis-cache-diagnostics">
+            <?php _e( 'Copy diagnostics to clipboard', 'redis-cache' ); ?>
+        </button>
+        <span class="success hidden" aria-hidden="true"><?php _e( 'Copied!' ); ?></span>
+    </span>
 </p>